refactor: move call to the Geonames API to its own module
The call to the Geonames API has been moved to a new geonames-api.js file
to declutter the main server.js
- Create geonames-api.js
- Import and configure geonamesAPI object in server.js
- Remove unused endpoint /location in server.js
